--- twine-dev/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "twine-dev"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-repository.workspace = true
-readme.workspace = true
-
-[lib]
-name = "twine_dev"
-
-[dependencies]
-twine-core = { version = "0.2.0", path = "../twine-core" }
-serde = { version = "1.0", features = ["derive"] }
-tokio = { version = "1", features = ["rt-multi-thread"] }
-warp = "0.3"
