+++ twine-examples/Cargo.toml
@@ -2,8 +2,8 @@
 name = "twine-examples"
 version.workspace = true
 edition.workspace = true
-license.workspace = true
 authors.workspace = true
+license.workspace = true
 repository.workspace = true
 readme.workspace = true
 description = "Example applications for Twine, a Rust framework for functional and composable system modeling."
@@ -13,7 +13,6 @@ keywords = ["twine", "framework", "examples", "demo"]
 serde = { version = "1.0", features = ["derive"] }
 twine-core = { path = "../twine-core" }
 twine-components = { path = "../twine-components" }
-twine-dev = {path = "../twine-dev"}
 twine-macros = { path = "../twine-macros" }
 twine-plot = { path = "../twine-plot" }
 twine-thermo = { path = "../twine-thermo" }
