+++ twine-plot/Cargo.toml
@@ -2,8 +2,8 @@
 name = "twine-plot"
 version.workspace = true
 edition.workspace = true
-license.workspace = true
 authors.workspace = true
+license.workspace = true
 repository.workspace = true
 readme.workspace = true
 description = "Plotting utilities for Twine, a Rust framework for functional and composable system modeling."
